Separate react typings to its own file.
